style : 아이콘 수정
