+++ src/components/common/Header/Header2.tsx
@@ -1,11 +1,10 @@
 'use client';
 
-import { ArrowLeftIcon } from '@/app/_components/icons';
 import { Button } from '@/components/ui/button';
 import { HOME } from '@/constants/pathname';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
-import { HomeIcon } from '../Icons';
+import { ArrowLeftIcon, HomeIcon } from '../Icons';
 
 const Header2 = () => {
   const router = useRouter();
